requirements.txt
coloredlogs
tqdm
pyyaml
tensorboardx
matplotlib